lvm: non-functional changes
Split into storage_lvm.go and storage_lvm_utils.go.
